requirements.txt
fastapi
MarkupSafe
multidict
openai
openpyxl
platformdirs
pydantic
pydantic-settings
python-multipart
requests
starlette
uvicorn
virtualenv
Jinja2
Markdown
MarkupSafe
beautifulsoup4
python-dotenv
